Sensor/Actuator

+++ src/critter/Actuator.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 public interface Actuator
 {
-	public void act (List<Double> input_, float tpf_);
+	public void act (List<Float> input_, float tpf_);
 	public int size ();
 }
